% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/StatComp20083R.R
\name{benchmarks}
\alias{benchmarks}
\title{Benchmark R and Rcpp functions.}
\description{
The functions used in my function (It's a function of my undergraduate paper)
}
\examples{
\dontrun{
n=100 
m=5    
max=1
min=0
i=1   
fangcha=0.01 
X=matrix(-1,nrow=n,ncol=m)
Y=rep(-1,n);U=Y;V=Y
while(Y[n]==-1)
{
  X[i,]=runif(m,min=min,max=max)
  gt=function(x){X[i,1]+X[i,2]*x+X[i,3]*x^2+X[i,4]*x^3+X[i,5]*x^4}
  y=as.numeric(integrate(gt,0,1)[1])+rnorm(1,mean=0,sd=fangcha)
  u=runif(1,min=-0.5,max=1.15);v=runif(1,min=1.15,max=2.8)
  if(u<=y&&y<=v){Y[i]=y;U[i]=u;V[i]=v;i=i+1}
}
rm(y)
Gn=creat(Y,U,V)
h1=seq(0.1,2,0.1)
h=CV(X,Y,U,V,h1)
h1=seq(max(0.01,h-0.1),h+0.1,0.01)
h=CV(X,Y,U,V,h1)
h1=seq(max(0.001,h-0.01),h+0.01,0.001)
h=CV(X,Y,U,V,h1)
h1=seq(max(0.0001,h-0.001),h+0.001,0.0001)
h=CV(X,Y,U,V,h1)
}
}
