% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/StatComp20083R_HW4.R
\name{benchmarks4}
\alias{benchmarks4}
\title{Benchmark4 R and Rcpp functions.}
\description{
The functions used in homework 4
}
\examples{
\dontrun{
HW4_1_1()
HW4_1_2()
HW4_2_1()
HW4_2_2()
HW4_2_3()
HW4_3_1()
HW4_3_2()
}
}
