% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/StatComp20083R_HW6.R
\name{benchmarks6}
\alias{benchmarks6}
\title{Benchmark6 R and Rcpp functions.}
\description{
The functions used in homework 6
}
\examples{
\dontrun{
HW6_1()
HW6_2_1()
HW6_2_2()
HW6_2_3()
HW6_2_4()
HW6_2_5()
}
}
