% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/StatComp20083R_HW8.R
\name{benchmarks8}
\alias{benchmarks8}
\title{Benchmark8 R and Rcpp functions.}
\description{
The functions used in homework 7
}
\examples{
\dontrun{
HW8_1()
HW8_2_1(mtcars)
HW8_2_2(mtcars)
HW8_3_1()
HW8_3_2()
HW8_4_1(mtcars, faithful)
HW8_4_2(mtcars, faithful)
}
}
