% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/StatComp20083R_HW9.R
\name{benchmarks9}
\alias{benchmarks9}
\title{Benchmark9 R and Rcpp functions.}
\description{
The functions used in homework 9
}
\examples{
\dontrun{
N=2000
sigma=c(.05,.5,2,16)

x0=25
rw1=rw.Metropolis(sigma[1],x0,N)
rw2=rw.Metropolis(sigma[2],x0,N)
rw3=rw.Metropolis(sigma[3],x0,N)
rw4=rw.Metropolis(sigma[4],x0,N)

crw1=crw(sigma[1],x0,N)
crw2=crw(sigma[2],x0,N)
crw3=crw(sigma[3],x0,N)
crw4=crw(sigma[4],x0,N)

qqplot(rw1$x,crw1$x)
qqplot(rw2$x,crw2$x)
qqplot(rw3$x,crw3$x)
qqplot(rw4$x,crw4$x)

tm1 <- microbenchmark::microbenchmark(
rw3 = rw.Metropolis(2,25,2000),
crw3 = crw(2,25,2000)
)
print(summary(tm1)[,c(1,3,5,6)])
}
}
