% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/StatComp20083R.R
\name{creat}
\alias{creat}
\title{Deal with the impact of truncation data on Y}
\usage{
creat(Y, U, V)
}
\arguments{
\item{Y}{a respond variable.}

\item{U}{double truncated about y.}

\item{V}{double truncated about y.}
}
\value{
a function
}
\description{
Deal with the impact of truncation data on Y
}
\examples{
\dontrun{
Y=c(1,1,1)
U=c(0.1,0.1,0.1)
V=c(2,2,2)
Gn=creat(Y,U,V,n)
}
}
