% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/StatComp20083R_HW1.R
\name{HW1_2}
\alias{HW1_2}
\title{The question 2 of homework 1}
\usage{
HW1_2()
}
\value{
nothing, but can get a picture.
}
\description{
The question 2 of homework 1
}
\examples{
\dontrun{
HW1_2()
}
}
