% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/StatComp20083R_HW1.R
\name{HW1_4}
\alias{HW1_4}
\title{The question 4 of homework 1}
\usage{
HW1_4()
}
\value{
nothing, but can get a picture.
}
\description{
The question 4 of homework 1
}
\examples{
\dontrun{
HW1_4()
}
}
