% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/StatComp20083R_HW4.R
\name{HW4_1_1}
\alias{HW4_1_1}
\title{the question 1.1 of homework (Beta)}
\usage{
HW4_1_1()
}
\value{
a data.frame
}
\description{
the question 1.1 of homework (Beta)
}
\examples{
\dontrun{
HW4_1_1()
}
}
