% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/StatComp20083R_HW4.R
\name{HW4_1_2}
\alias{HW4_1_2}
\title{the question 1.2 of homework (t(v))}
\usage{
HW4_1_2()
}
\value{
a data.frame
}
\description{
the question 1.2 of homework (t(v))
}
\examples{
\dontrun{
HW4_1_2()
}
}
