% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/StatComp20083R_HW4.R
\name{HW4_2_1}
\alias{HW4_2_1}
\title{the question 2.1 of homework(the Standard normal and the variance is the same)}
\usage{
HW4_2_1()
}
\value{
a data.frame
}
\description{
the question 2.1 of homework(the Standard normal and the variance is the same)
}
\examples{
\dontrun{
HW4_2_1()
}
}
