% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/StatComp20083R_HW4.R
\name{HW4_2_2}
\alias{HW4_2_2}
\title{the question 2.2 of homework(the Standard normal and the variance is not the same)}
\usage{
HW4_2_2()
}
\value{
a data.frame
}
\description{
the question 2.2 of homework(the Standard normal and the variance is not the same)
}
\examples{
\dontrun{
HW4_2_2()
}
}
