% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/StatComp20083R_HW4.R
\name{HW4_2_3}
\alias{HW4_2_3}
\title{the question 2.3 of homework(T-distribution(n=10) and the variance is the same)}
\usage{
HW4_2_3()
}
\value{
a data.frame
}
\description{
the question 2.3 of homework(T-distribution(n=10) and the variance is the same)
}
\examples{
\dontrun{
HW4_2_3()
}
}
