% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/StatComp20083R_HW4.R
\name{HW4_3_1}
\alias{HW4_3_1}
\title{the question 3.1 of homework(Standard normal)}
\usage{
HW4_3_1()
}
\value{
a data.frame
}
\description{
the question 3.1 of homework(Standard normal)
}
\examples{
\dontrun{
HW4_3_1()
}
}
