% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/StatComp20083R_HW4.R
\name{HW4_3_2}
\alias{HW4_3_2}
\title{the question 3.2 of homework(a special distribution)}
\usage{
HW4_3_2()
}
\value{
a picture
}
\description{
the question 3.2 of homework(a special distribution)
}
\examples{
\dontrun{
HW4_3_2()
}
}
