% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/StatComp20083R_HW5.R
\name{HW5_1}
\alias{HW5_1}
\title{the question 1 of homework}
\usage{
HW5_1(law)
}
\arguments{
\item{law}{a data frame (use bootstrap law)}
}
\value{
two number about bias.jack and se.jack
}
\description{
the question 1 of homework
}
\examples{
\dontrun{
HW5_1(law)
}
}
