% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/StatComp20083R_HW5.R
\name{HW5_2_1}
\alias{HW5_2_1}
\title{the question 2.1 of homework (The standard bootstrap CI based on standard normal)}
\usage{
HW5_2_1()
}
\value{
some number
}
\description{
the question 2.1 of homework (The standard bootstrap CI based on standard normal)
}
\examples{
\dontrun{
HW5_2_1()
}
}
