% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/StatComp20083R_HW5.R
\name{HW5_2_2}
\alias{HW5_2_2}
\title{the question 2.2 of homework (The standard bootstrap CI based on basic)}
\usage{
HW5_2_2()
}
\value{
some number
}
\description{
the question 2.2 of homework (The standard bootstrap CI based on basic)
}
\examples{
\dontrun{
HW5_2_2()
}
}
