% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/StatComp20083R_HW5.R
\name{HW5_2_4}
\alias{HW5_2_4}
\title{the question 2.4 of homework (The standard bootstrap CI based on BCa)}
\usage{
HW5_2_4()
}
\value{
some number
}
\description{
the question 2.4 of homework (The standard bootstrap CI based on BCa)
}
\examples{
\dontrun{
HW5_2_4()
}
}
