% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/StatComp20083R_HW5.R
\name{HW5_3}
\alias{HW5_3}
\title{the question 3 of homework}
\usage{
HW5_3(scor)
}
\arguments{
\item{scor}{a data frame (use bootstrap scor)}
}
\value{
some number
}
\description{
the question 3 of homework
}
\examples{
\dontrun{
HW5_3(scor)
}
}
