% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/StatComp20083R_HW5.R
\name{HW5_4}
\alias{HW5_4}
\title{the question 4 of homework}
\usage{
HW5_4(ironslag)
}
\arguments{
\item{ironslag}{a data frame (use DAAG ironslag)}
}
\value{
some number
}
\description{
the question 4 of homework
}
\examples{
\dontrun{
HW5_4()
}
}
