% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/StatComp20083R_HW6.R
\name{HW6_2_1}
\alias{HW6_2_1}
\title{the question 2.1 of homework(a normal distribution with a mean of 0 and a variance of 1 and 1.5)}
\usage{
HW6_2_1()
}
\value{
some number
}
\description{
the question 2.1 of homework(a normal distribution with a mean of 0 and a variance of 1 and 1.5)
}
\examples{
\dontrun{
HW6_2_1()
}
}
