% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/StatComp20083R_HW6.R
\name{HW6_2_2}
\alias{HW6_2_2}
\title{the question 2.2 of homework(a normal distribution with a mean of 0 and 1 and a variance of 1)}
\usage{
HW6_2_2()
}
\value{
some number
}
\description{
the question 2.2 of homework(a normal distribution with a mean of 0 and 1 and a variance of 1)
}
\examples{
\dontrun{
HW6_2_2()
}
}
