% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/StatComp20083R_HW6.R
\name{HW6_2_3}
\alias{HW6_2_3}
\title{the question 2.3 of homework(a t distribution with df=1)}
\usage{
HW6_2_3()
}
\value{
some number
}
\description{
the question 2.3 of homework(a t distribution with df=1)
}
\examples{
\dontrun{
HW6_2_3()
}
}
