% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/StatComp20083R_HW6.R
\name{HW6_2_4}
\alias{HW6_2_4}
\title{the question 2.4 of homework(A particular distribution)}
\usage{
HW6_2_4()
}
\value{
some number
}
\description{
the question 2.4 of homework(A particular distribution)
}
\examples{
\dontrun{
HW6_2_4()
}
}
