% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/StatComp20083R_HW6.R
\name{HW6_2_5}
\alias{HW6_2_5}
\title{the question 2.5 of homework(a normal distribution with a mean of 0 and a variance of 1, and the sample size of x is 20 and the sample size of y is 40)}
\usage{
HW6_2_5()
}
\value{
some number
}
\description{
the question 2.5 of homework(a normal distribution with a mean of 0 and a variance of 1, and the sample size of x is 20 and the sample size of y is 40)
}
\examples{
\dontrun{
HW6_2_5()
}
}
