% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/StatComp20083R_HW8.R
\name{HW8_2_1}
\alias{HW8_2_1}
\title{the question 2.1 of homework(for())}
\usage{
HW8_2_1(mtcars)
}
\arguments{
\item{mtcars}{a data frame (use datasets mtcars)}
}
\description{
the question 2.1 of homework(for())
}
\examples{
\dontrun{
HW8_2_1()
}
}
