% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/StatComp20083R_HW8.R
\name{HW8_4_1}
\alias{HW8_4_1}
\title{the question 4.1 of homework}
\usage{
HW8_4_1()

HW8_4_1()
}
\description{
the question 4.1 of homework

the question 4.2 of homework
}
\examples{
\dontrun{
HW8_4_1()
}
\dontrun{
HW8_4_2()
}
}
