% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/StatComp20083R_HW5.R
\name{benchmarks5}
\alias{benchmarks5}
\title{Benchmark5 R and Rcpp functions.}
\description{
The functions used in homework 5
}
\examples{
\dontrun{
HW5_1()
HW5_2_1()
HW5_2_2()
HW5_2_3()
HW5_2_4()
HW5_3()
HW5_4()
}
}
