% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/StatComp20083R_HW8.R
\name{HW8_4_1}
\alias{HW8_4_1}
\title{the question 4.1 of homework}
\usage{
HW8_4_1(mtcars, faithful)
}
\arguments{
\item{mtcars}{a data frame (use datasets mtcars)}

\item{faithful}{a data frame (use datasets faithful)}
}
\description{
the question 4.1 of homework
}
\examples{
\dontrun{
HW8_4_1()
}
}
